US4671 - Strong Auth
Updated with error handling for the 401 and repopulating the question. Did some layout changes to make Strong Auth look more like the comps. Added a custom radio button selector and am now extending not logged in activity for strong auth instead of roboactivity. Had to update other classes that where using not logged in activity since it now implements ErrorUI