US4631 Session Timeout
Now intercepting touches to determine timeout. Once timed out the user is sent home. On resume of the app if the user is timed out they will also be sent home. Also updated the log in activity to remove the launch mode of single task. Single task was causing the fragments to not resume.