US5472 Sliding drawer navigation
Updated the UI for the navigation. Added a divider as well as a footer and image for external links.
Updated the logic and added a new option for the menu in order to allow an on click to be passed instead of a fragment only.
Moved some logic around and did some refactoring of navigation item.
Integrated all of this with the new integration model for card and bank.