ID: DE2694 Name: Account Summary: App not displaying all returned Accounts
Root Cause:
Not allowing to add accounts to a groups if they did not match type. This was left from a previous bug fix where accounts weren't grouped correctly.

Tested:
Verified all accounts that are associated with signed in user are downloaded and grouped correctly
Tested in portrait and landscape orientations