DE2985 - Review Payments -Scheduled Payments screen has column labels Date/Payee/Amount along with text when all scheduled payments are deleted  - Android Only (QC# 89921)
Tested:
Verified header titles are not shown when user deletes the last transaction
Verified header titles are not shown if user navigates to review payments and there are no transactions
Verified header titles are now shown if user navigates from a tab that has transactions to a tab without transactions.
Verified headers are shown when user navigates from schedule to completed where scheduled has no transactions and completed does
Verified headers are shown when user navigates to review payments and there are transactions
Tested with orientation change with and without headers.