DE6492 - Hitting "Login" and then another button backs user out of login.
Created an InputEnablerListener which is used to disable all button on click listeners when login process begins.  The on click listeners are re-enabled through the completion listener of the  login service call.  Tested on both S3 4.1.1 and LG 2.3.4.  Test Case: Type in credentials into username and password field.  Hit the "Login" button and quickly tapped on other UI elements, for example the "Contact Us" button.